+++ src/Client/ServiceExtensions.cs
@@ -1,8 +1,8 @@
-﻿using System.Diagnostics.CodeAnalysis;
+﻿using SmartComponents.LocalEmbeddings;
+using System.Diagnostics.CodeAnalysis;
 using Tavenem.Wiki;
 using Tavenem.Wiki.Blazor;
 using Tavenem.Wiki.Blazor.Client;
-using Tavenem.Wiki.Blazor.Services.Search;
 
 namespace Microsoft.Extensions.DependencyInjection;
 
@@ -27,8 +27,7 @@ public static class ServiceExtensions
     /// An <see cref="ISearchClient"/> instance.
     /// </para>
     /// <para>
-    /// If omitted, an instance of <see cref="DefaultSearchClient"/> will be used. Note: the
-    /// default client is not recommended for production use.
+    /// If omitted the built-in search functionality will be used.
     /// </para>
     /// </param>
     /// <returns>The <see cref="IServiceCollection"/> instance.</returns>
@@ -39,6 +38,7 @@ public static class ServiceExtensions
         ISearchClient? searchClient = null)
     {
         services
+            .AddSingleton<LocalEmbedder>()
             .AddTavenemFramework()
             .AddScoped<WikiState>();
 
@@ -60,11 +60,7 @@ public static class ServiceExtensions
             services.AddScoped(_ => new WikiBlazorClientOptions());
         }
 
-        if (searchClient is null)
-        {
-            services.AddScoped<ISearchClient, DefaultSearchClient>();
-        }
-        else
+        if (searchClient is not null)
         {
             services.AddScoped(_ => searchClient);
         }
@@ -83,8 +79,7 @@ public static class ServiceExtensions
     /// The type of <see cref="ISearchClient"/> to register.
     /// </para>
     /// <para>
-    /// If omitted, <see cref="DefaultSearchClient"/> will be used. Note: the default client is
-    /// not recommended for production use.
+    /// If omitted the built-in search functionality will be used.
     /// </para>
     /// </param>
     /// <param name="wikiOptions">
@@ -102,6 +97,7 @@ public static class ServiceExtensions
         WikiBlazorClientOptions? wikiBlazorOptions = null)
     {
         services
+            .AddSingleton<LocalEmbedder>()
             .AddTavenemFramework()
             .AddScoped<WikiState>();
 
@@ -123,11 +119,7 @@ public static class ServiceExtensions
             services.AddScoped(_ => new WikiBlazorClientOptions());
         }
 
-        if (searchClientType is null)
-        {
-            services.AddScoped<ISearchClient, DefaultSearchClient>();
-        }
-        else
+        if (searchClientType is not null)
         {
             services.AddScoped(typeof(ISearchClient), searchClientType);
         }
@@ -152,8 +144,7 @@ public static class ServiceExtensions
     /// A function which provides an <see cref="ISearchClient"/> instance.
     /// </para>
     /// <para>
-    /// If omitted, an instance of <see cref="DefaultSearchClient"/> will be used. Note: the
-    /// default client is not recommended for production use.
+    /// If omitted the built-in search functionality will be used.
     /// </para>
     /// </param>
     /// <returns>The <see cref="IServiceCollection"/> instance.</returns>
@@ -164,16 +155,13 @@ public static class ServiceExtensions
         Func<IServiceProvider, ISearchClient>? searchClientBuilder = null)
     {
         services
+            .AddSingleton<LocalEmbedder>()
             .AddTavenemFramework()
             .AddScoped<WikiState>()
             .AddScoped(wikiOptionsBuilder)
             .AddScoped(wikiBlazorOptionsBuilder);
 
-        if (searchClientBuilder is null)
-        {
-            services.AddScoped<ISearchClient, DefaultSearchClient>();
-        }
-        else
+        if (searchClientBuilder is not null)
         {
             services.AddScoped(searchClientBuilder);
         }