+++ src/Client/Services/ISearchClient.cs
@@ -1,4 +1,4 @@
-﻿using Tavenem.Wiki.Blazor.Services.Search;
+﻿using Tavenem.Wiki.Queries;
 
 namespace Tavenem.Wiki.Blazor;
 
@@ -11,11 +11,11 @@ public interface ISearchClient
     /// Search for wiki content which matches the given search criteria.
     /// </summary>
     /// <param name="request">
-    /// An <see cref="ISearchRequest"/> instance with search criteria.
+    /// A <see cref="SearchRequest"/> instance with search criteria.
     /// </param>
     /// <param name="user">
     /// The <see cref="IWikiUser"/> making the request.
     /// </param>
-    /// <returns>An <see cref="ISearchResult"/> instance with search results.</returns>
-    Task<ISearchResult> SearchAsync(ISearchRequest request, IWikiUser? user);
+    /// <returns>A <see cref="SearchResult"/> instance with search results.</returns>
+    Task<SearchResult> SearchAsync(SearchRequest request, IWikiUser? user);
 }